Add a ros node to drive the car using learned parameters
