+++ ros_ws/src/autonomous/q_learning/scripts/car.py
@@ -1,5 +1,4 @@
 import rospy
-from std_msgs.msg import String, Empty
 from gazebo_msgs.msg import ModelState, ModelStates
 from sensor_msgs.msg import LaserScan
 from drive_msgs.msg import drive_param
@@ -9,16 +8,7 @@ import math
 
 import torch
 
-from simulation_tools.track import track
 from simulation_tools.reset_car import Point
-
-def _crash_callback(_):
-    for function in _crash_callback_functions:
-        function()
-
-
-def register_crash_callback(function):
-    _crash_callback_functions.append(function)
 
 
 def _model_state_callback(message):
@@ -59,17 +49,10 @@ def get_scan(sample_count, device):
     values = [v if not math.isinf(v) and not math.isnan(v) else 100 for v in values]
     return torch.tensor(values, device=device)
 
-def get_distance_to_track_center():
-    track_position = track.localize(current_position)
-    return track_position.distance_to_center
-
 
 current_position = None
 laser_scan = None
 
-_crash_callback_functions = []
-
-rospy.Subscriber("/crash", Empty, _crash_callback)
 rospy.Subscriber("/gazebo/model_states", ModelStates, _model_state_callback)
 rospy.Subscriber("/scan", LaserScan, _laser_callback)
 drive_parameters_publisher = rospy.Publisher(
